requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.2.5
autopep8==1.4.4
autosub==0.4.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cvxopt==1.2.3
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
dominate==2.5.2
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.27.2
h5py==2.10.0
httplib2==0.18.1
idna==2.8
imageio==2.5.0
intel-openmp==2019.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.17
jedi==0.13.2
Jinja2==2.10
jsonpatch==1.26
jsonpointer==2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.2.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mkl==2019.0
mpmath==1.1.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.3
nose==1.3.7
notebook==5.7.4
numpy==1.16.2
oauthlib==3.1.0
opencv-contrib-python==4.1.0.25
opt-einsum==3.2.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.2
pickleshare==0.7.5
Pillow==5.4.1
progressbar==2.5
progressbar2==3.51.4
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pygame==1.9.4
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
pysrt==1.1.2
python-dateutil==2.7.5
python-nmap==0.6.1
python-utils==2.4.0
pytz==2018.9
PyWavelets==1.0.3
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scapy==2.4.3
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sympy==1.3
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
torch==1.6.0
torchfile==0.1.0
torchfunc==0.2.0
torchvision==0.7.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.1
uritemplate==3.0.1
urllib3==1.25.6
virtualenv==16.6.1
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.4.2
wrapt==1.11.1
